% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countK.R
\name{countK}
\alias{countK}
\title{count the occurrences of a provided keyword}
\usage{
countK(var1, key, data)
}
\arguments{
\item{key}{the string or number to match}

\item{data}{specify a dataset for the variables x and y}

\item{x}{a variable}
}
\value{
countK returns the number of times a key occurs in x
}
\description{
count the occurrences of a provided keyword
}
\examples{
countK(Gender,"M",data=StudentSurvey,)
IQRK(Smoke,"Y")
}
\keyword{bRth}
