% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottyK.r
\name{dottyK}
\alias{dottyK}
\title{Make a dotplot-style histogram from small sample}
\usage{
dottyK(var1, data, xlab, main)
}
\arguments{
\item{data}{the dataset from which x comes}

\item{xlab}{x axis label}

\item{main}{main title}

\item{x}{a quantitative variable}
}
\value{
dotplot
}
\description{
Make a dotplot-style histogram from small sample
}
\examples{
dottyK(Height,data=StudentSurvey) 
dottyK(Height,data=StudentSurvey,main="Heights from Student Survey Data")
}
\keyword{bRth}
