% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivenumK.R
\name{fivenumK}
\alias{fivenumK}
\title{Compute the five number summary of a quantitative variable}
\usage{
fivenumK(var1, data, na.rm = FALSE)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{select whether to omit NA, default is FALSE}

\item{x}{a variable}
}
\value{
five number summary
}
\description{
Compute the five number summary of a quantitative variable
}
\examples{
fivenumK(Height,data=StudentSurvey)
fivenumK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
