% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histK.r
\name{histK}
\alias{histK}
\title{Draw a histogram}
\usage{
histK(var1, data, ...)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{x}{a variable}
}
\value{
histK is a wrapper for base::hist
   that adds the functionality of specifying data=<dataset>
   in the input arguments
}
\description{
Draw a histogram
}
\examples{
meanK(Height,data=StudentSurvey)
meanK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
