% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRK.R
\name{IQRK}
\alias{IQRK}
\title{Compute the inter quartile range (IQR) of a quantitative variable}
\usage{
IQRK(var1, data, na.rm = FALSE, type = 7)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{select whether to omit NA, default is FALSE}

\item{x}{a variable}
}
\value{
IQR
}
\description{
Compute the inter quartile range (IQR) of a quantitative variable
}
\examples{
IQRK(Height,data=StudentSurvey)
IQRK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
