% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanK.R
\name{meanK}
\alias{meanK}
\title{Compute the mean of a quantitative variable}
\usage{
meanK(var1, data, na.rm = FALSE)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{select whether to omit NA, default is FALSE}

\item{x}{a variable}
}
\value{
mean value
}
\description{
Compute the mean of a quantitative variable
}
\examples{
meanK(Height,data=StudentSurvey)
meanK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
