% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianK.R
\name{medianK}
\alias{medianK}
\title{Compute the median of a quantitative variable}
\usage{
medianK(var1, data, na.rm = FALSE)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{select whether to omit NA, default is FALSE}

\item{x}{a variable}
}
\value{
standard deviation
}
\description{
Compute the median of a quantitative variable
}
\examples{
medianK(Height,data=StudentSurvey)
medianK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
