% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdK.R
\name{sdK}
\alias{sdK}
\title{Compute the standard deviation of a quantitative variable}
\usage{
sdK(var1, data, na.rm = FALSE)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{select whether to omit NA, default is FALSE}

\item{x}{a variable}
}
\value{
sdK is a wrapper for base::sd
   that adds the functionality of specifying data=<dataset>
   in the input arguments
}
\description{
Compute the standard deviation of a quantitative variable
}
\examples{
sdK(Height,data=StudentSurvey)
sdK(Height,data=StudentSurvey,na.rm=TRUE)
}
\keyword{bRth}
