% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableK.R
\name{tableK}
\alias{tableK}
\title{Create a table from categorical variable(s)}
\usage{
tableK(var1, var2 = NULL, data)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{x}{a variable or a formula x~y}

\item{y}{a variable (set to NULL if x is a formula)}
}
\value{
a table
}
\description{
Create a table from categorical variable(s)
}
\examples{
tableK(Gender,data=StudentSurvey)
table(Gender~Smoke, data=StudentSurvey)
table(StudentSurvey$Gender,StudentSurvey$Smoke)
}
\keyword{bRth}
