% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corK.R
\name{corK}
\alias{corK}
\title{Compute correlation from quantitative variable(s)}
\usage{
corK(arg1, arg2 = NULL, data, na.rm = FALSE)
}
\arguments{
\item{data}{specify a dataset for the variables x and y}

\item{na.rm}{determine whether to omit NA,  default is FALSE}

\item{x}{a variable or a formula x~y}

\item{y}{a variable (set to NULL if x is a formula)}
}
\value{
correlation
}
\description{
Compute correlation from quantitative variable(s)
}
\examples{
corK(Height,Weight,data=StudentSurvey)
corK(Height~Weight, data=StudentSurvey)
corK(StudentSurvey$Gender,StudentSurvey$Smoke,na.rm=TRUE)
}
\keyword{bRth}
